Rearrangement of struct molData
The reason for this is to separate the pointers which the parallel threads need to write to from the remainder of the molData
attributes, which are read-only within the crucial code blocks. I put three of the attributes in question, 'jbar' 'phot' and
'vfac', into a separate structure, which is now declared, allocated, written to and freed within the code lines which will
eventually be within the parallel block. The attribute 'ds' does not depend on species and so was just declared as a pointer to
double outside of any struct. I also renamed it to halfFirstDs to be a bit more descriptive.

The run time and output values appear to be unaffected.