Added a layer of randomness in photon and raytrace
Again this is done to prepare for the parallelisation, when we will need separate RNGs for the separate threads. Note that we
will still get repeatable results when the TEST flag in the makefile is set. The output is now slightly different than the
previous commit produced, because in effect different random seeds are in use both in the random selection of initial photon
directions and the random dither of starting rays in the raytracing section. The present output however is expected to be
identical to that produced by single-thread running of the parallelized code.