New configuration struct.
The new struct is typedefd as configInfo. It contains all the information which was
previously stored in inputPars: the difference is that inputPars has now been trimmed to
contain _only_ those parameters which are settable by the user in the model.c file. Any
non-user-settable values should now be added to the config struct.

The advantages of this are as follows.

	- We will usually want more config parameters than user-settable ones. The separation
          leaves it clearer which things the user needs to (or can) set.

	- The separation makes it easier to check for and exclude impossible combinations of
          parameters.

	- We can adopt new names (for clarity) for config parameters without bothering the
          user with a changed interface.

Two small unrelated additional changes:

	- A new macro MAX_NIMAGES is defined in lime.h and used in the initial malloc of struct image
          instead of MAX_NSPECIES as before.

	- Fixed the order of functions in lime.h so they are all (instead of just mostly, as before)
          alphabetical.