Rearranged grid-point selection code.
Code that was in grid.c and weights.c has been put in a new module, random_points.c.
This is in preparation for the introduction of a new grid-point-selection algorithm.
The module weights.c is no longer compiled, since it is now obsolete.

Additional changes connected with the above:
	- Added a new (optional) user parameter 'samplingAlgorithm'.
	- Added external variables modelRadiusSquared, densityNormalizer,
          numCollisionPartners.
	- Added macro DENSITY_POWER.
	- New struct locusType.