Added new grid-point-selection algorithm.
The new algorithm calls the function randomsViaTree() recursively in a 2^D-tree
divide-and-conquer approach. The aim of this is to allow more accurate
representation of models for which the density varies strongly across the model,
without the severe time penalties that straightforward use of the rejection method
would entail. (The previous algorithm uses rejection method, but it 'flattens' the
density function in order to avoid long run times, which has the effect that
high-density regions are undersampled.) The algorithm assesses the density gradients
over a sub-volume of the model space, and if those gradients are judged too large
for quick results from the rejection method, the sub-volume is divided into 8
smaller sub-volumes and the same function is called for each one of these. The
division is performed by cutting the volume approximately in halves (at the 0.5
point plus some random dither) along each of the (3) axes of the coordinate space.

Additional changes connected with the above:
	- New macros:
		N_TREE_RANDOMS
		MAX_RECURSION
		MAX_N_TRIALS_TREE
		TREE_DITHER