Added some write points
Here I have added some places where the grid is written to file. Everything seems to work
ok, and the output values look sensible, but I have not yet done any exhaustive tests of
the values in the fits files.

Some new elements of the struct inputPars have been necessary. The default values of these
cause no fits output to be written. In fact there should be no changes in the way the
software functions with the standard model.c file.

I also added a new function printMessage() to messages.c and some lines in main.c which
print the number of threads used when this is more than 1.

The fits-read functionality remains unused. Adding the ability to read the grid from file
should perhaps wait on discussion among the developers about the future of the pregrid etc
machinery.