Added new structs.
This is the first of many commits I plan for an eventual pull request which will provide
LIME with a smoother raytracing capability. This added functionality requires quite a
lot of new code and will also involve some cleaning up of existing code.

The smoother raytracing algorithm works by linear interpolation of level populations
inside the Delaunay cells of the model grid. This first commit lays the groundwork for
reading and using the cell information by defining several new structs to allow
convenient storage and manipulation of this data.

Structs added were: cell, pop2, gridInterp, gAuxType, intersectType, faceType and
triangle2D. None of these are yet active so as yet they cause no change in the task
function.

Also added in this commit is a new user parameter traceRayAlgorithm. This will be used
to invoke the new raytracing algorithm; at present it has no effect.

Finally, I have commented out the function template for FastExp() in lime.h and added
the variable type of the input argument of greetings_parallel() in messages.c just to
avoid the warning messages that gcc issues now on compilation. This is just for present
convenience, I plan to make these minor fixes the subject of a separate PR.