New functions sourceFunc_*_raytrace().
These duplicate the functionality of sourceFunc_cont() and sourceFunc_line() but are
made necessary, or at least desirable, by the present poor way in which grid data is
stored in memory structures. When this is fixed (which it will hopefully be eventually),
the *_raytrace versions can be merged into the previous ones.