Rearranged if-block in traceray().
There is a test near the start of the function to see whether the ray will intersect the
model sphere at all; if so, the following large block of ~110 statements (containing the
entire functionality of the function) is entered. It is obviously much less spendthrift
of indents to simply exit the function if the ray does _not_ pass through the model
sphere. The code has been rearranged to implement this change. All the statements which
were formerly within the block have been unindented by 2 spaces; there are no further
changes in the present commit.