Smarter call to velocity in raytrace
This improves the speed of raytracing by implementing a suggestion made by T Lunttila in
the discussion of issue #95. In the present master branch, the user-specified function
velocity() is called for every channel/line combination judged to have a significant
line amplitude; but of course the velocity depends only on position in the model. The
function is actually called N times (currently N=10) at N spatially separated points
along the path of the ray from entry to exit of a particular Voronoi cell. In the
present commit, the N velocity values have been precalculated just once per cell. This
speeds raytracing up on my desktop machine from 25 sec to 11 sec.

In addition, I have
  - cleaned up some divisions by int within velocityspline2();

  - tightened up the function interface to velocityspline2();

  - removed the raytrace_1.4 code, because (i) his is available elsewhere, (ii) I'll try
    to distil it into a PR soon, and (iii) it was generating errors now due to an
    altered traceray() interface.