Smarter call to sourceFunc_cont
The call to this function has been moved outside the loop over channels in traceray().
It is only necessary to call it once per channel, then just keep the returned values.
The change does not speed things up much, possibly because sourceFunc_cont() is not
actually doing many FPOs.

I also introduced some variables to make traceray() a leeetle more self-documenting and
a leeetle less obscure. Lot of work still to go in this direction in LIME as a whole.
:-/

A final change was removing the unused argument 'vfac' from sourceFunc_pol().