Remove surplus call to sortangles
The previous commit allows the two calls to sortangle() (one before the loop commencement and the 2nd at the
end of the loop block) to be merged into a single call at the start of the loop block. Ditto with setting
'there'. This has a small effect on the output image, even with the random seed frozen via setting the test
flag, because the order of access of the random numbers has changed.