Pointless line loop entry for continuum.
Formerly, if img[im].doline==0, nlinetot was set to 1. This is pointless however. The
loop in traceray() over lines was entered, but for each line the contribution is only
calculated if img[im].doline!=0. Now I have set nlinetot to zero in this case and removed
the test on img[im].doline in the line-loop in traceray().