Abolished lineCount()
The motivation behind this change is to replace the single loops over the total line
tally which occur in photon.c and raytrace.c with double nested loops over
respectively the number of radiating species (par.nSpecies) and the number of lines
for that species. This in turn makes it much easier to implement blending in both the
photon() and getjbar() routines without hacks to the object which will store the
information about blended lines.

The changes include:
  - Removal of lineCount(), all calls to it, plus the counta, countb thingies.
  - Replacing of all loops over nlinetot by two nested loops as described above.
*NOTE* that I have not yet altered the indent level inside these loops, so as to make
the job of comparison easier.
  - I commented out lineBlend() pro tem since I am going to rewrite all this code
anyway.
  - I removed raytrace_1_4() to avoid the tedium of updating the traceray() interface
in this unused code.