Added 9 new functions.
These new functions are mostly intended to replace molinit() and kappa(). Splitting these functions into chunks of
separate functionality is not only a good idea from the point of view of code clarity and maintainability, it is
absolutely necessary to allow some of the traps and bugs of the old code to be removed.

New functions added in this commit (but not yet used in the remainder of the code):
  - aux.c:checkUserDensWeights(). This functionality could have been placed in aux.c:parseInput() but for one
reason: that it depends on the number of densities, and if par->doPregrid or par->restart, this number is not
known at the time of call of parseInput().

  - grid.c:gridLineInit(): contains some of the grid-specific stuff which was in the old molinit().

  - grid.c:calcGridMolDensities(): this replaces the block of code beginning at line 240 of the old molinit.c. The
algorithm has been greatly expanded, not only to cater for the new user-settable inpars attributes collPartIds,
nMolWeights and dustWeights, but to close off some of the traps the previous code left. Some additional warnings
have been added. Note also that calculation of the .nmol quantities (via the 2nd of the two functions) now is done
the same way for LTE as for non-LTE.

  - grid.c:calcGridDustOpacity(): contains the grid-specific functionality of the old kappa().

  - grid.c:calcGridCollRates(): functionality is the same as the block of code in lines 259-287 of the old
molinit.c.

  - molinit.c:calcMolCMBs(): contains the molecule-specific functionality of the old kappa().

  - molinit.c:readMolData(): does what it says on the box.

  - molinit.c:setUpDensityAux(): if the user does not supply par.collPartIds and/or par.nMolWeights, the task
tries its best to mimic the behaviour of Lime before the present patch. (Note that complete replication of the old
behaviour would also mean replicating the errors and traps in the old algorithm, which of course we do not wish to
do.)

  - molinit.c:assignMolCollPartsToDensities().