Main structural changes to make LIME callable from ARTIST
Moved the initialization of the input parameters, allocation, release and initialization
of the output fits files from aux.c to the routines initParImg and freeParImg in main.c.

Moved the main part of the main program to the new routine run in main.c.

These changes allow the new run routine to be used as an interface to LIME from external
programs (e.g. ARTIST). While for stand-alone LIME the routines initParImg and freeParImg
will be called for the initialization of the input parameters and output fit images, external
programs using LIME must do this initialization before calling run.