Made silent a global variable instead of a constant defined in header file
The Python interface of LIME has functions setSilent and getSilent to allow the
user control over whether status information is accessible from Python or not.
Therefore, silent has to be a variable instead of a constant.

While this is not critical for ARTIST, there seems to be no argument against
making silent a variable, and it would allow the LIME interface for ARTIST to
remain unchaged.