New interface for IO of grid data.
This is a substantial commit, but it is hard to see how it could have been easily broken
into smaller ones. Here I have reworked the previous facility for IO to and from FITS-format
files. There is another, generic IO layer now implemented in a new module gridio.c, with
FITS relegated to a choie of file format. Hopefully this will make it easier if anybody else
wants to write an interface to HDF5 or a DB or whatever.

Some detail on the changes:
  - New pars par->gridInFile, par->gridOutFiles[] (user-settable) and par->writeGridAtStage[], par->dataStageI (not intended to be user settable).
  - New function grid.c:mallocAndSetDefaultGrid().
  - grid.c:buildGrid() rewritten and renamed readOrBuildGrid().
  - main.c changed to include the fix to #130.