More detail now printed if par.collPartIds not supplied
I moved collpartnames to the module header of molinit.c so all the functions can see it.

In the case that the user doesn't set par.collPartIds, a warning is given that the task will
have to guess the identities of the collison partners. I have added lines which say what the
result of those guesses was.