Added header file for inpars
Following advice from @rschaaf I have created a separate header file inpars.h just to define
the struct for the user-input parameters.

I also included macros in both lime.h and inpars.h to prevent multiple inclusion of these
files.

The final change was to the interface of openSocket(). This had the inpars struct as one of
its arguments, the other being the species ID, but it turns out that all it needs is the
name of the moldatfile for that species. Thus I replaced the former two arguements by a
single char * moldatfile.