Added 1.4 raytrace algorithm
The algorithm in 1.4 was much faster than the one in 1.5 because it solved the radiative
transfer equation for a smaller number of rays, essentially corresponding to the locations
of the grid points when projected onto the image plane. This has now been imported into the
present merged raytracing code. I changed the 1.4 code in two respects:
  - The smoothing step is no longer done.
  - Instead of the somewhat arcane image-plane interpolation that is done between sparse
rays in 1.4, a straightforward linear interpolation is performed.