Added another loop layer in grid point select.
The algorithm to choose grid point locations is sort of a quasi rejection method. Instead of looping more or less endlessly, generating
both a new random 'rejection tester' R as well as a new trial point location at each iteration, which the classic rejection method does,
the Lime algorithm chooses a single R between 0 and 1 and just loops over trial positions. This can lead to endless loops if there happens
to be no place in the model for which fracDensity is returned with a value higher than R. One could move the generation of R inside the
inner loop, but this might have unforeseen effects on the choice algorithm, which contains other non-rejection features. Instead I have
included another loop layer outside the choice of R which is intended to be a kind of safety layer to turn endless loops into merely long
ones and to generate a warning message if the situation is encountered.