Grid i/o made more flexible
There are two main changes here, although there are a lot of changed lines of code.

  - A generic i/o layer has been added, with choice of FITS just one of the (currently the
only) possibilities for the format of the disk files. Hopefully this will make it easier for
people to add interfaces to HDF5 or whatever at a later date.

  - The information which can legally be read from or written to the files has now been made
much more flexible. When reading the file, if a particular element of the disk struct, say
density, was not supplied, the task will automatically seek to fill in these missing values via
the standard route. The completion state of the grid struct is now indicated by bit flags
rather than a single 'stage' integer.