Changed the way image units are specified to negate needless repetitions of raytracing.
Currently the user specifies an image unit for each image produced and then raytracing is performed
for each of these images. However, if the user wants different unit output using the same set of
image parameters then this is entirely unecessary as all the information for every image unit is
computed through one iteration of the raytracing process. Therefore I have changed the way units are
specified to a delimited string so that the user can specify multiple unit outputs for each image to
be outputted. I have also added in an automatic suffixing of the image unit to the image filename via
a fitsFilename() function in writefits.c to make the output units easily identifiable.