Fixes data race in raythrucells.
The function extractFace() was written to return a pointer to a faceType struct. The return value was copied from a local variable which I had made static to avoid compiler warnings. However of course static variables are shared over threads, so this was causing a data race condition. I fixed this by having the function return a faceType struct directly rather than a pointer to one. That allows me to avoid the static declaration inside the function.