Moved calculation of number of species in grid2fits to before grid is malloc'd.
Before, in grid2fits.c the grid was malloc'd initially with a call to mallocAndSetDefaultGrid()
using nSpecies=0. However, this meant that (*gp).[i].mol was never malloc'd so that later when
abundances were read in it would cause a segfault. This is a simple fix that moves nSpecies
determination to before the mallocAndSetDefaultGrid() call and passes this as an argument so that
(*gp).[i].mol is malloc'd correctly if appropriate.